added test for discharge figure

+++ tests/test_utils.py
@@ -115,8 +115,10 @@ class Test_VisualiseOutput:
         
         # 2. Set expectations
         output_dir = TestUtils.get_local_test_file('cases/case_02_compound/output/figures/roughness')
-        shutil.rmtree(output_dir)
+        
+        if output_dir.is_dir(): shutil.rmtree(output_dir)
         os.mkdir(output_dir)
+        
         output_file = TestUtils.get_local_test_file('cases/case_02_compound/output/figures/roughness/roughness_longitudinal_channel1.png')
         
         # 3. Run test
@@ -139,7 +141,7 @@ class Test_VisualiseOutput:
 
         # 2. Set expectations
         output_dir = TestUtils.get_local_test_file('cases/case_02_compound/output/figures/cross_sections')
-        shutil.rmtree(output_dir)
+        if output_dir.is_dir(): shutil.rmtree(output_dir)
         os.mkdir(output_dir)
         output_file = TestUtils.get_local_test_file('cases/case_02_compound/output/figures/cross_sections/channel1_125.000.png')
         
@@ -214,6 +216,28 @@ class Test_Compare1D2D:
         # 3. Run test
         try:
             plotter.figure_longitudinal(route=['BR', "PK", "IJ"], stat="last25")
+        except Exception as e:
+            pytest.fail("No exception expected, but thrown: {}".format(str(e)))
+
+        # 4. Verify expectations
+        assert output_file.is_file()
+
+    def test_figure_discharge(self):
+
+        # 1. Set up initial test data 
+        project_config = TestUtils.get_local_test_file('compare1d2d/rijn-j22_6-v1a2/sobek-rijn-j22.ini')
+        project = Project(project_config)
+        plotter = Compare1D2D(project=project,
+                    start_time=datetime(year=2000, month=1, day=5))
+        
+        # 2. Set expectations
+        # this file should exist
+        output_file = TestUtils.get_local_test_file('compare1d2d/rijn-j22_6-v1a2/output/figures/discharge/Pannerdensche Kop.png')
+
+        # 3. Run test
+        try:
+            plotter.figure_compare_discharge_at_stations(stations=['WL_869.00', 'PK_869.00'], 
+                                             title="Pannerdensche Kop")
         except Exception as e:
             pytest.fail("No exception expected, but thrown: {}".format(str(e)))
